feat: 지도 이동 기능 구현
